Creates a new version of Set that handles any SELECT query.
This changes the focus of this set to allow handling of any kind of
select query, without any dependency on a DBObject\Item.  In doing
so, I've removed the ability to add new object from outside, as there
is no reasonable control for what they would look like.

Created some testing for this.  Very basic stuff, just to make sure
the workflow was happening.

I hope to later utilize this as a way to also incorporate WITH and
UNION type queries to run on here.