Puts in place proper handling for primary key fields for an insert
An empty primary key value on an insert usually means the field is an
auto incrementer, so don't include it in the list of fields getting
values.

Also, we may have a primary key that is manually assigned, so I can't
just assume every primary key is auto incrementing.  If the field has
a value, then go ahead and try and put it into the fields being
inserted.