Expands the ability of an Item Set to use multiple SQL engines
Many times a simple SELECT simply isn't enough.  Also, DBSql readily
provides a great way to create WITH and UNION queries that I wanted
to make use of.  I also wanted to be able to push raw SQL directly in
without having to use DBSql at all if I don't want to.

To do this, there are now 4 distinct engines that can be used.  To
determine which one should be used, the last engine accessed is the
one to run.