Provide a way to fetch an item based on the primary key value
After using this object for a while, I'm finding that I'm over using the
find() method for fetching an item based on the primary key.  I didn't
want to change the nature of get(), since there is value in being able
to pull something based on the index in the data set.

I realize getPk() is an awful name.  I wanted something short, that
at least kind of explains what it does.